refactor: remove ModelColumns.php trait
